fix: consolidate lifecycle-control files for pkg-file-count CI gate
Merge durable lifecycle handlers, mapper errors, servertests, service
tests, and fixture runtime lifecycle coverage into existing package
owners so all affected packages stay within the 15-file limit.
