+++ pkg/api/servertests/server_durable_session_lifecycle_control_test.go
@@ -9,6 +9,7 @@ import (
 	"net/http/httptest"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 	"time"
 
@@ -738,4 +739,41 @@ func postFactorySessionRetryDispatchRaw(
 		return nil, errors.New(errResp.Message)
 	}
 	return resp, nil
+}
+func TestRealBackendFactorySessionRoutes_LifecycleControlsAreImplemented(t *testing.T) {
+	projectRoot := setupAPIRuntimeWorkflowFixture(t, "busy-loop.workflow.js", "busy-loop")
+	service := factorysessionexecution.NewJavaScriptRuntimeService(factorysessionexecution.JavaScriptRuntimeServiceConfig{
+		ProjectRoot: projectRoot,
+	})
+	srv := newAPITestServer(&testutil.MockFactory{DurableExecutionService: service})
+	server := httptest.NewServer(srv.Handler())
+	defer server.Close()
+
+	started := postDurableAsyncStart(t, server.URL, factoryapi.FactorySessionExecutionRequest{
+		RequestId: "req-api-runtime-lifecycle-slice-001",
+		Source: factoryapi.FactorySessionExecutionSource{
+			Kind:         factoryapi.FactorySessionExecutionSourceKindWorkflowName,
+			WorkflowName: strPtr("busy-loop"),
+		},
+	})
+	sessionPath := "/factory-sessions/" + started.SessionId
+
+	resp, err := http.Post(server.URL+sessionPath+"/cancel", "application/json", strings.NewReader(`{}`))
+	if err != nil {
+		t.Fatalf("POST %s/cancel: %v", sessionPath, err)
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusAccepted {
+		t.Fatalf("POST %s/cancel status = %d, want 202: %s", sessionPath, resp.StatusCode, readBody(t, resp))
+	}
+	var control factoryapi.FactorySessionLifecycleControlResponse
+	if err := json.NewDecoder(resp.Body).Decode(&control); err != nil {
+		t.Fatalf("decode lifecycle control response: %v", err)
+	}
+	if control.Operation != factoryapi.FactorySessionLifecycleControlKindCancel {
+		t.Fatalf("operation = %q, want CANCEL", control.Operation)
+	}
+	if control.Outcome != factoryapi.FactorySessionLifecycleControlOutcomeAccepted {
+		t.Fatalf("outcome = %q, want ACCEPTED", control.Outcome)
+	}
 }
